requirements.txt
python-dotenv
requests
mcp[cli]
